Moved undertow config to spring app entry config
