+++ Spring-server/src/main/java/ikpi63holding/postclient/Application.java
@@ -1,8 +1,16 @@
 package ikpi63holding.postclient;
 
+import io.undertow.UndertowOptions;
+import io.undertow.servlet.api.SecurityConstraint;
+import io.undertow.servlet.api.SecurityInfo;
+import io.undertow.servlet.api.TransportGuaranteeType;
+import io.undertow.servlet.api.WebResourceCollection;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.embedded.undertow.UndertowServletWebServerFactory;
+import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 
@@ -10,6 +18,33 @@ import org.springframework.context.annotation.Bean;
 public class Application {
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
+    }
+
+    @Value("${server.port.http}")
+    int httpPort;
+
+    @Value("${server.port}")
+    int httpsPort;
+
+    @Value("${server.address}")
+    String address;
+
+    @Bean
+    public ServletWebServerFactory servletContainer() {
+        UndertowServletWebServerFactory undertow = new UndertowServletWebServerFactory();
+        undertow.addBuilderCustomizers(builder -> {
+            builder.addHttpListener(httpPort, address);
+            builder.setServerOption(UndertowOptions.ENABLE_HTTP2, true);
+        });
+        undertow.addDeploymentInfoCustomizers(deploymentInfo -> {
+            deploymentInfo.addSecurityConstraint(new SecurityConstraint()
+                    .addWebResourceCollection(new WebResourceCollection()
+                            .addUrlPattern("/*"))
+                    .setTransportGuaranteeType(TransportGuaranteeType.CONFIDENTIAL)
+                    .setEmptyRoleSemantic(SecurityInfo.EmptyRoleSemantic.PERMIT))
+                    .setConfidentialPortManager(exchange -> httpsPort);
+        });
+        return undertow;
     }
 
     @Bean
